requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AWS SDK (uses IAM role - no keys needed!)
boto3==1.34.34
botocore==1.34.34

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Google AI
google-generativeai==0.3.2

# Image Processing (optional)
Pillow==10.2.0
opencv-python==4.9.0.80

# Environment Variables
python-dotenv==1.0.1

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development Tools (optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
